moved the subjects.completed column to quarter.completed.

+++ src/main/webapp/app/model/subject/SubjectsModel.js
@@ -67,7 +67,7 @@ Ext.define('MySchool.model.subject.SubjectsModel', {
         },
         {
             name: 'qtrGrade',
-            type: 'double'
+            type: 'float'
         },
         {
             name: 'qtrGradeType',
